pyproject.toml
[project]
name = "bohrium-mcp"
version = "0.1.0"
description = "Bohrium MCP implementation including client, host and server components"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # 核心依赖
    "mcp>=1.9.0",
    "requests>=2.32.3",
    "openai>=1.12.0",
    "python-dotenv>=1.0.0",
    # 其他依赖（从 requirements.txt 迁移过来）
    "annotated-types>=0.7.0",
    "anyio>=4.9.0",
    "httpx>=0.28.1",
    "pydantic>=2.11.5",
    "python-multipart>=0.0.20",
    "starlette>=0.46.2",
    "uvicorn>=0.34.2",
    "httpx-sse>=0.4.0",
    "jiter>=0.10.0",
    "pytz>=2024.1"
]

[project.optional-dependencies]
dev = [
    "ipykernel>=6.29.5",
    "ipython>=9.2.0",
    "debugpy>=1.8.14",
]

test = [
    "pytest>=7.0.0",
]

[[project.authors]]
name = "hzt"
email = "zhentinghng@gmail.com"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
openapi-mcp-server = "openapi_mcp_server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/cli", "src/mcp_host", "src/openapi_mcp_server"]
